refactor: Move XML helper functions to separate xml_utils module
I've refactored the XML helper functions (find_tag_value_start,
get_tag_string, set_tag_string) out of sysprod.c/h and into a
new, separate module (xml_utils.c and xml_utils.h).

Changes include:
- Created `source/xml_utils.h` with the declarations for the three
  XML helper functions.
- Created `source/xml_utils.c` with the implementations of these
  functions.
- Removed the XML helper function implementations from `source/sysprod.c`
  and added an include for `xml_utils.h`.
- Removed the XML helper function declarations from `source/sysprod.h`.
- Updated `source/test_sysprod.c` to include `xml_utils.h` for the
  declarations of the functions it tests.

This refactoring improves code organization, separates general XML
parsing/manipulation logic from sysprod-specific concerns, and makes
the XML utilities more easily reusable in other parts of the project
if needed.

Note: The Makefile will need to be updated to compile and link
`source/xml_utils.c` and to handle new header dependencies.
